adjusting prioritizer to fit new taskList.h
